don't overwrite gcpm.conf

+++ install.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 
 cp bin/gcpm /usr/bin/
-cp etc/gcpm.conf /etc/
+[ -f /etc/gcpm.conf ] || cp etc/gcpm.conf /etc/
 cp system/gcpm.service /etc/systemd/system/
